+++ src/VBox/Storage/QED.cpp
@@ -1,4 +1,4 @@
-/* $Id: QED.cpp 48743 2013-09-27 18:19:03Z alexander.eichner@oracle.com $ */
+/* $Id: QED.cpp 48851 2013-10-03 20:02:34Z knut.osmundsen@oracle.com $ */
 /** @file
  * QED - QED Disk image.
  */
@@ -2525,7 +2525,7 @@ static int qedSetParentFilename(void *pBackendData, const char *pszParentFilenam
 
                     Assert((offData & UINT32_MAX) == offData);
                     pImage->offBackingFilename = (uint32_t)offData;
-                    pImage->cbBackingFilename  = strlen(pszParentFilename);
+                    pImage->cbBackingFilename  = (uint32_t)strlen(pszParentFilename);
                     rc = vdIfIoIntFileSetSize(pImage->pIfIo, pImage->pStorage,
                                               offData + pImage->cbCluster);
                 }
